Major changes on Analyzer and Reade - mltthread
Program data - added varialbe for storing queue_max_length, two condition variables for signaling full and empty queue, added pointer to Analyzer data, fixed bug with not initializing cpu_usage mutex
Added a function for waiting in a thred for cnd_t signal with wimeout of 1 sec
Changed analyzer function to a multithreaded one
Changed Reader function to a multithreded one

+++ src/Printer.c
@@ -29,7 +29,7 @@ int Printer_thread(void *pdv)
         mtx_lock(&pd->mtx_cpu_usage);
 
         printw("cpu usage %d: %.2f %%\n", test++, pd->cpu_usage[0] * 100);
-        for (size_t i = 1; i < pd->core_cnt + 1; i++)
+        for (int i = 1; i < pd->core_cnt + 1; i++)
         {
             printw(" cpu%d : %.2f %%\n", i - 1, pd->cpu_usage[i] * 100);
         }
